add observation area
