+++ src/main/java/com/gosnack/snack_order_api/model/ItemModel.java
@@ -9,9 +9,6 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 
-
-import java.time.LocalDateTime;
-
 @Entity
 @Table(name = "items")
 public class ItemModel {
@@ -28,7 +25,6 @@ public class ItemModel {
 
     private String itemName;
     private double itemPrice;
-    private LocalDateTime orderTime = LocalDateTime.now();
 
     public Long getItemId() {
         return itemId;
@@ -62,11 +58,4 @@ public class ItemModel {
         this.itemPrice = itemPrice;
     }
 
-    public LocalDateTime getOrderTime() {
-        return orderTime;
-    }
-
-    public void setOrderTime(LocalDateTime orderTime) {
-        this.orderTime = orderTime;
-    }
 }
