+++ src/main/java/com/gosnack/snack_order_api/utils/OrderStatus.java
@@ -2,7 +2,9 @@ package com.gosnack.snack_order_api.utils;
 
 public enum OrderStatus {
     PEDIDO_ACEITO,
+    PEDIDO_CANCELADO,
     PEDIDO_EM_PREPARO,
+    PEDIDO_PRONTO,
     PEDIDO_SAIU_PARA_ENTREGA,
     PEDIDO_ENTREGUE
 }
